+++ src/tasks/dotfiles.rs
@@ -3,6 +3,8 @@ use std::path::Path;
 use utils;
 
 pub fn sync() {
+    println!("pkg: dotfiles: syncing ...");
+
     let target = utils::env::home_dir().join(Path::new(".dotfiles"));
 
     if utils::git::has_git() && utils::git::path_is_git_repository(&target) {