+++ src/tasks/hyper.rs
@@ -3,6 +3,8 @@ use std::path::Path;
 use utils;
 
 pub fn sync() {
+    println!("pkg: hyper: syncing ...");
+
     let src = utils::env::home_dir().join(Path::new(".dotfiles/config/hyper.js"));
     let dest = utils::env::home_dir().join(Path::new(".hyper.js"));