+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* pkg: vim: Windows: _vimrc instead of .vimrc
+
 ## [0.6.0] - 2018-03-10
 
 ### Added