+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Windows: ensure ~/bin exists in PATH
 
-* pkg: [skaffold](https://github.com/GoogleCloudPlatform/skaffold): `sync` downloads into ~/bin
+* pkg: [shfmt](https://github.com/mvdan/sh): `sync` (into ~/bin) and `update`
 
-* pkg: [skaffold](https://github.com/GoogleCloudPlatform/skaffold): `update` downloads into ~/bin when version is outdated
+* pkg: [skaffold](https://github.com/GoogleCloudPlatform/skaffold): `sync` (into ~/bin) and `update`
 
 ## [0.6.1] - 2018-03-14